server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Igor Cardoso <igorgcardoso7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
django-sequences = "^2.7"
django-cors-headers = "^3.13.0"
python-decouple = "^3.6"
psycopg2-binary = "^2.9.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
